+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/memory/MemoryFragment.kt
@@ -23,8 +23,8 @@ import com.woowacourse.staccato.presentation.memory.viewmodel.MemoryViewModel
 import com.woowacourse.staccato.presentation.memory.viewmodel.MemoryViewModelFactory
 import com.woowacourse.staccato.presentation.memoryupdate.MemoryUpdateActivity
 import com.woowacourse.staccato.presentation.moment.MomentFragment.Companion.MOMENT_ID_KEY
+import com.woowacourse.staccato.presentation.momentcreation.VisitCreationActivity
 import com.woowacourse.staccato.presentation.util.showToast
-import com.woowacourse.staccato.presentation.visitcreation.VisitCreationActivity
 
 class MemoryFragment :
     BindingFragment<FragmentMemoryBinding>(R.layout.fragment_memory),
