+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/memorycreation/MemoryCreationActivity.kt
@@ -20,8 +20,8 @@ import com.woowacourse.staccato.presentation.common.PhotoAttachFragment
 import com.woowacourse.staccato.presentation.memory.MemoryFragment.Companion.MEMORY_ID_KEY
 import com.woowacourse.staccato.presentation.memorycreation.viewmodel.MemoryCreationViewModel
 import com.woowacourse.staccato.presentation.memorycreation.viewmodel.MemoryCreationViewModelFactory
+import com.woowacourse.staccato.presentation.momentcreation.OnUrisSelectedListener
 import com.woowacourse.staccato.presentation.util.showToast
-import com.woowacourse.staccato.presentation.visitcreation.OnUrisSelectedListener
 
 class MemoryCreationActivity : BindingActivity<ActivityMemoryCreationBinding>(), MemoryCreationHandler, OnUrisSelectedListener {
     override val layoutResourceId = R.layout.activity_memory_creation
