+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/momentcreation/dialog/MemorySelectionHandler.kt
@@ -1,0 +1,7 @@
+package com.woowacourse.staccato.presentation.momentcreation.dialog
+
+import com.woowacourse.staccato.presentation.momentcreation.model.MomentMemoryUiModel
+
+fun interface MemorySelectionHandler {
+    fun onConfirmClicked(memoryUiModel: MomentMemoryUiModel)
+}
