+++ android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/momentcreation/dialog/VisitedAtSelectionFragment.kt
@@ -1,4 +1,4 @@
-package com.woowacourse.staccato.presentation.visitcreation.dialog
+package com.woowacourse.staccato.presentation.momentcreation.dialog
 
 import android.os.Bundle
 import android.view.LayoutInflater
