--- android/Staccato_AN/app/src/main/java/com/woowacourse/staccato/presentation/visitcreation/dialog/MemorySelectionHandler.kt
@@ -1,7 +1,0 @@
-package com.woowacourse.staccato.presentation.visitcreation.dialog
-
-import com.woowacourse.staccato.presentation.visitcreation.model.MomentMemoryUiModel
-
-fun interface MemorySelectionHandler {
-    fun onConfirmClicked(memoryUiModel: MomentMemoryUiModel)
-}
