+++ android/Staccato_AN/app/src/main/res/layout/activity_visit_creation.xml
@@ -10,11 +10,11 @@
 
         <variable
             name="viewModel"
-            type="com.woowacourse.staccato.presentation.visitcreation.viewmodel.VisitCreationViewModel" />
+            type="com.woowacourse.staccato.presentation.momentcreation.viewmodel.VisitCreationViewModel" />
 
         <variable
             name="visitCreationHandler"
-            type="com.woowacourse.staccato.presentation.visitcreation.VisitCreationHandler" />
+            type="com.woowacourse.staccato.presentation.momentcreation.VisitCreationHandler" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -41,7 +41,7 @@
             <androidx.constraintlayout.widget.ConstraintLayout
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                tools:context=".presentation.visitcreation.VisitCreationActivity">
+                tools:context=".presentation.momentcreation.VisitCreationActivity">
 
                 <TextView
                     android:id="@+id/tv_visit_creation_place_name_title"
